Correct unsigned-signed comparison

+++ src/asmif.c
@@ -221,7 +221,7 @@ static int fixup_deferred_exprs(struct parse_data *pd)
         if (ce_eval(pd, ce->insn, NULL, ce, 0, sym_reloc_handler, &r->width, &result)) {
             result *= r->mult;
 
-            if (result != SEXTEND32(r->width, result)) {
+            if (result != (int32_t)SEXTEND32(r->width, result)) {
                 debug(0, "Expression resulting in value %#x is too large for "
                         "%d-bit signed immediate field", result, r->width);
                 rc |= 1;
